refactor: name of Octree namespace to Optimisation
