+++ src/Plugins/Primitives/Cylinder/Cylinder.cpp
@@ -129,9 +129,9 @@ Math::Vector3D Primitive::Cylinder::getNormal(const Math::Vector3D& hitPoint) co
     return Math::Vector3D (hitPoint.x() - _origin.x(), hitPoint.y() - _origin.y(), hitPoint.z() - _origin.z());
 }
 
-Octree::cubeCollider Primitive::Cylinder::getColliderBox() const
+Optimisation::cubeCollider Primitive::Cylinder::getColliderBox() const
 {
-    Octree::cubeCollider collider;
+    Optimisation::cubeCollider collider;
 
     collider.maxX.second = _origin.x() + _radius;
     collider.minX.second = _origin.x() - _radius;
